Main: Warnings
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@63161 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/include/GuestCtrlImplPrivate.h
@@ -822,7 +822,7 @@ public:
     void Dump(const char *pszFile);
 #endif
 
-    uint32_t GetOffset() { return m_offBuffer; }
+    size_t GetOffset() { return m_offBuffer; }
 
     size_t GetSize() { return m_cbUsed; }
